Support backslash based escaping for placeholders
The simpleParse method of parser, which get invoked if message contains
only placeholders, was not supporting escaped placeholders(\\$10)

The light weight versio of jquery.i18n when parser.js not included
also has the same fix now.

Add support for that and add tests.
Updated the documentation in README.md too

Fixes issue #74